requirements.txt
Django==1.10.6
django-discover-runner==1.0
django-extensions==1.7.7
djangorestframework==3.6.2
olefile==0.44
Pillow==4.0.0
PyMySQL==0.7.10
six==1.10.0
